:sparkles: 部分文件修改

+++ src/entities/weather-city.entity.ts
@@ -32,7 +32,7 @@ export class WeatherCity extends BaseEntity {
   address: string
 
   @Column({
-    type: 'decimal',
+    type: 'float',
     precision: 8,
     scale: 5,
     comment: '经度', // 浮点数，范围为-180~180，负数表示西经。使用 gcj02 国测局坐标系
@@ -40,7 +40,7 @@ export class WeatherCity extends BaseEntity {
   longitude: number
 
   @Column({
-    type: 'decimal',
+    type: 'float',
     precision: 8,
     scale: 5,
     comment: '纬度', // 浮点数，范围为-90~90，负数表示南纬。使用 gcj02 国测局坐标系
